1) std::begin,  std::end 代替 boost::begin, boost::end 2) OArchive::save_binary 和 IArchive::load_binary 改为 public 3) 去掉一些类型转换的警告
git-svn-id: svn://23.27.117.66/h2/trunk@223 54bad739-14b1-1b4c-9595-0bb2987a2747

+++ lv/lv/Stream/IBufferStream.hpp
@@ -39,7 +39,7 @@ namespace lv
 			if (size >= 0)
 			{
 				to_read = std::min<streamsize>(size, stream_size() - gpos_);
-				std::copy(buf_.begin() + gpos_, buf_.begin() + gpos_ + to_read, data);
+				std::copy(buf_.begin() + static_cast<size_t>(gpos_), buf_.begin() + static_cast<size_t>(gpos_ + to_read), data);
 				
 				gpos_ += to_read;
 			}
@@ -114,7 +114,7 @@ namespace lv
 					break;
 				}
 
-				unsigned char c = buf_[pos];
+				unsigned char c = buf_[static_cast<size_t>(pos)];
 				if (static_cast<int>(c) == delim)
 				{
 					break;
